Backend/app/requirements.txt
Flask==2.2.3
Pytest==7.2.1
flask-pymongo
flask_jwt_extended
Werkzeug==2.2.2
requests
flask_cors